presentations/test-the-web-forward/index.html

<!doctype html>  
<html lang="en">
	
	<head>
		<meta charset="utf-8">
		
		<title>W3C Test Writing Tutorials</title>

		<meta name="description" content="An easy to use CSS 3D slideshow tool for quickly creating good looking HTML presentations.">
		<meta name="author" content="Hakim El Hattab">
		
		<link href='http://fonts.googleapis.com/css?family=Lato:400,700,400italic,700italic' rel='stylesheet' type='text/css'>
		
		<link rel="stylesheet" href="css/reset.css">
		<link rel="stylesheet" href="css/main.css">
		<link rel="stylesheet" href="css/adobe.css">
		<link rel="stylesheet" href="lib/zenburn.css">
	</head>
	
	<body>
		
		<div id="reveal">
			<!-- Any section element inside of this container is displayed as a slide -->
			<div class="slides">
			    <script type="text/javascript">
  			    function computedStyle(element)
          	{
          	  document.getElementById("computedStyle").innerHTML = window.getComputedStyle(element).getPropertyValue("-webkit-transform");
          	}

          </script>
                <section>
                    <h2>Test The Web Forward</h2>
                    <img src="assets/images/W3C_logo.jpg">
                    <script>
    					// Delicously hacky. Look away.
    					if( navigator.userAgent.match( /(iPhone|iPad|iPod|Android)/i ) )
    					document.write( '<p style="color: rgba(0,0,0,0.3); text-shadow: none;">('+'Tap to navigate'+')</p>' );
    				</script>
                </section>

                <section id="Agenda">
                   <h2>Agenda</h2>
                   <ul>
                       <li>Setup</li>
                       <li>Ref Test Example</li>
                       <li>JavaScript Test Example</li>
                    </ul>
                </section>

                <section>
                    <h2>Setup</h2>
                    <h4>Complete all steps before proceeding</h4>
                    <br>
                    <ul>
                        <li>Step 1: Create a CSSWG Test Manager Account<br><a href="https://test.csswg.org/shepherd/register/" style="font-size:18px">https://test.csswg.org/shepherd/register/</a> 
                            <ul>
                                <li>Select the check box next to "Request Repository Write Access"</li>
                            </ul>
                        </li>
                        <li>Step 2: Install Mercurial<br><a href="http://wiki.csswg.org/tools/hg/install" style="font-size:18px">http://wiki.csswg.org/tools/hg/install</a></li>
                        <li>Step 3: Setup Mercurial Preferences<br><a href="http://wiki.csswg.org/tools/hg#setting-up-mercurial-preferences" style="font-size:20px">http://wiki.csswg.org/tools/hg#setting-up-mercurial-preferences</a></li>   
                        <li>Step 4: Clone the CSSWG Test Suite Repository<br><a href="http://wiki.csswg.org/tools/hg#initial-clone-of-the-central-repository" style="font-size:20px">http://wiki.csswg.org/tools/hg#initial-clone-of-the-central-repository</a></li>
                        <li>Step 5: Clone the Resources directory<br><a href="http://wiki.csswg.org/tools/hg#initial-clone-of-the-central-repository" style="font-size:20px">http://wiki.csswg.org/tools/hg#initial-clone-of-the-central-repository</a></li>
                    </ul>
                </section>
                
               	<section>
					<h2>Reftest Overview</h2>
					<br>
					<blockquote>
					<ul>
					    <li>A reftest consists of two files:</li>
					    <ul>
					        <li>Test file - Uses the feature under test</li>
					        <li>Reference file - An exact visual match without using feature under test</li>
						</ul>
						<li>Self-describing - Works both manually and with automation</li>
						<li>Cross-browser & cross-platform</li>
					</ul>
					</blockquote>
				</section>
				
				<section>
    				<h2>Directory structure</h2>
    				<section>
    				<br>
    				<ul>
    				    <li>CSSWG Repository</li>
    				    <ul>
    				        <li>test</li>
    				        <ul>
    				            <li>contributors</li>
						        <ul>
						            <li>ttwf
						            <ul>
						                <li>yourTestDir</li>
						                <ul>
						                    <li>incoming</li>
    						                <li>submitted</li>
						                </ul>
						            </ul>
						        </ul>
						    </ul>
						</ul>
						</li>
					</ul>
					</section>
					<section>
					    <img src="assets/images/directoryStructure.png"/>
					</section>
    			</section>
    			<section>
    				<h2>Sample Test Case</h2>
    				<section>
    					<br><br>
    					<p>Basic test for the transform property described in the CSS3 Transforms spec using the translate() function</p>
    					<br>
    					<br>
    					<p>Spec Links:</p>
    					<p style="font-size: 16pt">
    					    <a href="http://dev.w3.org/csswg/css3-transforms/#transform-property" target="top">http://dev.w3.org/csswg/css3-transforms/#transform-property</a>
    					    <br>
    					    <a href="http://dev.w3.org/csswg/css3-transforms/#two-d-transform-functions" target="top">http://dev.w3.org/csswg/css3-transforms/#two-d-transform-functions</a>
					    </p>
					</section>
					<section>
					    <img src="assets/images/specAssertion.png">
					</section>
	            </section>
				
			    <section>
			        <section>
			            <h2>Step 1 - OSX</h2>
			            <h5>Press down-arrow for Windows instructions</h5>
    			        <br>
        			    <h3>Update your local CSSWG test repository</h3>
        			    <p style="font-size:22px">Complete this step unless you just cloned the repository in the previous step</p>
        			    <br>
			        
    			        <p style="font-size:30px">From a terminal window, type the following:</p>
      				    <div class="console">
      				        <div class="consoleText">
          				        <br><br>
          			            <p>$ cd <span class="localPath">[path to repo]</span>/test</p>
          			            <p>$ hg pull -u</p>
          			            <p>$ ls -lR contributors/ttwf/samples</p>
          			        </div>
      			        </div>
      			        <br><br><br>
      			        <div class="bottomNote">
      			            <h5>
      			                Note: These files can also be downloaded from here:
      			                <br>
      			                <a href="assets/resources/ttwf-reftest-tutorial.zip">ttwf-reftest-tutorial.zip</a>
      			            </h5>
      			        </div>
  			        </section>  
  			        <section>
                        <h2>Step 1 - Windows</h2>
			            <h5>Press up-arrow for OSX instructions</h5>
    			        <br>
        			    <h3>Update your local CSSWG test repository</h3>
        			    <p style="font-size:22px">Complete this step unless you just cloned the repository in the previous step</p>
        			    <br>
			        
    			        <p style="font-size:30px">From a DOS command window, type the following:</p>
      				    <div class="console">
      				        <div class="dosText">
          				        <br><br>
          			            <p>> cd <span class="localPath">[path to repo]</span>\test</p>
          			            <p>> hg pull -u</p>
          			            <p>> dir contributors\ttwf\samples</p>
          			        </div>
      			        </div>
      			        <br><br><br>
      			        <div class="bottomNote">
      			            <h5>
      			                Note: These files can also be downloaded from here:
      			                <br>
      			                <a href="assets/resources/ttwf-reftest-tutorial.zip">ttwf-reftest-tutorial.zip</a>
      			            </h5>
      			        </div>                                                 
      			    </section>
                </section>
                
                <section>
                    <section>
    			         <h2>Step 2 - OSX</h2>
    			         <h5>Press down-arrow for Windows instructions</h5>
        			     <br>
            		     <h3>Create a directory for your new test</h3>
        			     <br>
          				 <div class="console">
          				     <div class="consoleText">
              				     <br><br><br>
              				     <p>$ cd <span class="localPath">[path to repo]</span>/test/contributors/ttwf/</p>
              				     <p>$ mkdir -p <span class="localPath">yourName</span>/incoming</p>
              				     <p>$ cd <span class="localPath">yourName</span>/incoming</p>
              			     </div>
          			     </div>
           		    </section>
           		    <section>
           		        <h2>Step 2 - Windows</h2>
           		        <h5>Press up-arrow for OSX instructions</h5>
        			    <br>
            		    <h3>Create a directory for your new test</h3>
        			    <br>
           				<div class="console">
           				    <div class="dosText">
               				     <br>
               				     <br>
               				     <br>
               				     <p>> cd <span class="localPath">[path to repo]</span>\test\contributors\ttwf</p>
               				     <p>> mkdir <span class="localPath">yourName</span>\incoming</p>
               				     <p>> cd <span class="localPath">yourName</span>\incoming</p>
               			    </div>
           			    </div>
           		    </section>
                </section>
                
                <section>
                    <section>
    			         <h2>Step 3 - OSX</h2>
    			         <h5>Press down-arrow for Windows instructions</h5>
        			     <br>
            		     <h3>Copy the starter files to your incoming directory</h3>
            		     <br>
          				 <div class="console">
          				     <div class="consoleText">
              				     <br>
              				     <br>
              				     <p><span class="localPath">From your incoming directory created in previous step</span></p>
              			         <p>$ cp -R ../../samples/* .</p>
              			     </div>
          			     </div>
                    </section>
                    <section>
                        <h2>Step 3 - Windows</h2>
    			         <h5>Press up-arrow for OSX instructions</h5>
        			     <br>
            		     <h3>Copy the starter files to your incoming directory</h3>
            		     <br>
          				 <div class="console">
          				     <div class="dosText">
              				     <br>
              				     <br>
              				     <p><span class="localPath">From your incoming directory created in previous step</span></p>
              			         <p>> xcopy /e ..\..\samples .</p>
              			     </div>
          			     </div>                        
                    </section>
                </section>
                
                <section>
    			     <h2>Step 4</h2>
    			     <h3>Open the <span style="color:red">ttwf-reftest-tutorial-001.html</span> file in a editor</h3>
    			     <img src="assets/images/testStarter.png">
                </section>
                
                <section>
    			     <h2>Step 5</h2>
    			     <br>
        		     <h3>Add the test metadata</h3>
    			     <div class="tab5">
    			        <ul>
      				        <li>Title</li>
      				        <li>Author</li>
      				        <li>Spec link(s)</li>
      				        <li>Reference file path</li>
      				        <li>Flags</li>
      				        <li>Assertion</li>
      				    </ul>
      				</div>
      				<div class="bottomNote">
  			            <h5>
  			                Note: Each metadata element is described in detail on the CSSWG Wiki<br>
  			                <a href="http://wiki.csswg.org/test/format#template-details">http://wiki.csswg.org/test/format#template-details</a>
  			            </h5>
  			        </div>
                </section>
                
                <section>
    			     <h2>Step 6</h2>
    			     <br>
        		     <h3>Add the transform attribute and test function to the SVG element</h3>
        		     <br>
        		     <div class="editor">
       				 <div class="editorText">
       				     <br><br>
       				     <p class="tab1 editorHtml">&lt;body&gt;
       				     <p class="tab2 editorHtml">&lt;svg&gt;</p>
       			         <p class="tab3 editorHtml">&lt;rect width=<span class="editorAttrVal">"100"</span>
       			             <span class="editorHtml">height=</span><span class="editorAttrVal">"100"</span>
       			             <span class="highlight"><span class="editorHtml">transform=</span><span class="editorAttrVal">"translate(50)"</span></span><span class="editorHtml">/&gt;</span>
       			         </p>
       			         <p class="tab2 editorHtml">&lt;/svg&gt;</p>
       			         <p class="tab1 editorHtml">&lt;/body&gt;
       			     </div>
       			     </div>
    		    </section>
    		    
                <section>
    			     <h2>Step 7</h2>
    			     <br>
        		     <h3>Create an element that is visible only upon failure</h3>
    			     <br>
         		     <div class="editor">
        			 <div class="editorText">
            		     <br>
            		     <p class="tab1 editorHtml">&lt;body&gt;
            		     <p class="tab2 editorHtml">&lt;svg&gt;</p>
            		     <p class="tab3 editorHtml"><span class="highlight">&lt;rect x=<span class="editorAttrVal">"51"</span>
              		         <span class="editorHtml">y=</span><span class="editorAttrVal">"1"<span>
              		         <span class="editorHtml">width=</span><span class="editorAttrVal">"98"</span>
              		         <span class="editorHtml">height=</span><span class="editorAttrVal">"98"</span><span class="editorHtml">/&gt;</span></span>
              		     </p>
              		     <p class="tab3 editorHtml">&lt;rect width=<span class="editorAttrVal">"100"</span>
            		         <span class="editorHtml">height=</span><span class="editorAttrVal">"100"</span>
            		         <span class="editorHtml">transform=</span><span class="editorAttrVal">"translate(50)"<span><span class="editorHtml">/&gt;</span>
            		     </p>
            		     <p class="tab2 editorHtml">&lt;/svg&gt;</p>
            			 <p class="tab1 editorHtml">&lt;/body&gt;
            		 </div>
        			 </div>
                </section>
                
                <section>
    			     <h2>Step 8</h2>
    			     <br>
        		     <h3>Add green & red fill colors to denote passing & failing conditions</h3>
        		     <br>
        		     <div class="editor">
         			 <div class="editorText">
         			     <p class="tab1 editorHtml">&lt;body&gt;
             		     <p class="tab2 editorHtml">&lt;svg&gt;</p>
             		     <p class="tab3 editorHtml">&lt;rect width=<span class="editorHtml">x=</span><span class="editorAttrVal">"51"<span>
               		         <span class="editorHtml">y=</span><span class="editorAttrVal">"1"<span>
               		         <span class="editorHtml">width=</span><span class="editorAttrVal">"98"</span>
               		         <span class="editorHtml">height=</span><span class="editorAttrVal">"98"</span>
               		         <span class="highlight"><span class="editorHtml">fill=</span><span class="editorAttrVal">"red"</span></span><span class="editorHtml">/&gt;</span>
               		     </p>
             		     <p class="tab3 editorHtml">&lt;rect width=<span class="editorAttrVal">"100"</span>
             		         <span class="editorHtml">height=</span><span class="editorAttrVal">"100"</span>
             		         <span class="editorHtml">transform=</span><span class="editorAttrVal">"translate(50)"<span>
             		         <span class="editorHtml"><span class="highlight">fill=<span class="editorAttrVal">"green"</span></span><span class="editorHtml">/&gt;</span>
             		     </p>
             		     <p class="tab2 editorHtml">&lt;/svg&gt;</p>
             			 <p class="tab1 editorHtml">&lt;/body&gt;
             		 </div>
         			 </div>
    			</section>
    			
                <section>
    			     <h2>Step 9</h2>
    			     <br>
        		     <h3>Add a statement to the test page to make this a self-describing test</h3>
        		     <br>
        		     <div class="editor">
         			 <div class="editorText">
         			     <p class="tab1 editorHtml">&lt;body&gt;
             		     <p class="tab2 editorHtml"><span class="highlight">&lt;p&gt;<span class="editorHtmlTxt">The test passes if there is a green square and no red.</span><span class="editorHtml">&lt;/p&gt;</span></span>
                         <p class="tab2 editorHtml">&lt;svg&gt;</p>
             		     <p class="tab3 editorHtml">&lt;rect x=</span><span class="editorAttrVal">"51"<span>
               		         <span class="editorHtml">y=</span><span class="editorAttrVal">"1"<span>
               		         <span class="editorHtml">width=</span><span class="editorAttrVal">"98"</span>
               		         <span class="editorHtml">height=</span><span class="editorAttrVal">"98"</span>
               		         <span class="editorHtml">fill=</span><span class="editorAttrVal">"red"</span><span class="editorHtml">/&gt;</span>
               		     </p>
             		     <p class="tab3 editorHtml">&lt;rect width=<span class="editorAttrVal">"100"</span>
             		         <span class="editorHtml">height=</span><span class="editorAttrVal">"100"</span>
             		         <span class="editorHtml">transform=</span><span class="editorAttrVal">"translate(50)"<span>
             		         <span class="editorHtml">fill=</span><span class="editorAttrVal">"green"</span><span class="editorHtml">/&gt;</span>
             		     </p>
             		     <p class="tab2 editorHtml">&lt;/svg&gt;</p>
             			 <p class="tab1 editorHtml">&lt;/body&gt;
             		 </div>
         			 </div>
    			</section>
    			
                <section>
    			     <h2>Step 10</h2>
    			     <br>
        		     <h3>Open the <span style="color:red">reftests/ttwf-reftest-tutorial-ref.html</span> file in an editor</h3>
        		     <img src="assets/images/refStarter.png">
    		    </section>
    		    
    		    <section>
    			     <h2>Step 11</h2>
    			     <br>
        		     <h3>Add the reference metadata</h3>
    			     <div class="tab5">
    			        <ul>
      				        <li>Title</li>
      				        <li>Author</li>
      				        <li>Flags</li>
      				    </ul>
      				</div>
                </section>
                
                <section>
        		     <h2>Step 12</h2>
        		     <br>
            	     <h3>Create and position the reference element so it matches the test file exactly in a passing condition</h3>
            	     <br>
            	        <div class="editor">
        				<div class="editorText">
        				     <br><br>
        				     <p class="tab1 editorHtml">&lt;body&gt;
        				     <p class="tab2 editorHtml">&lt;svg&gt;</p>
        			         <p class="tab3 editorHtml"><span class="highlight">&lt;rect x=<span class="editorAttrVal">"50"</span>
        			             <span class="editorHtml">y=</span><span class="editorAttrVal">"0"<span>
                                 <span class="editorHtml">width=</span><span class="editorAttrVal">"100"<span>
                                 <span class="editorHtml">height=</span><span class="editorAttrVal">"100"</span>
        			             <span class="editorHtml">fill=</span><span class="editorAttrVal">"green"</span><span class="editorHtml">/&gt;</span></span>
        			         </p>
        			         <p class="tab2 editorHtml">&lt;/svg&gt;</p>
        			         <p class="tab1 editorHtml">&lt;/body&gt;
        			    </div>
        			    </div>
                </section>
                
                <section>
        		     <h2>Step 13</h2>
        		     <br>
            	     <h3>Add the self-describing statement here, too</h3>
            	     <br>
            	     <div class="editor">
     				 <div class="editorText">
     				     <br><br>
     				     <p class="tab1 editorHtml">&lt;body&gt;
     				     <p class="tab2 editorHtml"><span class="highlight">&lt;p&gt;<span class="editorHtmlTxt">The test passes if there is a green square and no red.</span><span class="editorHtml">&lt;/p&gt;</span></span>
                         <p class="tab2 editorHtml">&lt;svg&gt;</p>
     			         <p class="tab3 editorHtml">&lt;rect x=<span class="editorAttrVal">"50"</span>
     			             <span class="editorHtml">y=</span><span class="editorAttrVal">"0"<span>
                              <span class="editorHtml">width=</span><span class="editorAttrVal">"100"<span>
                              <span class="editorHtml">height=</span><span class="editorAttrVal">"100"</span>
     			             <span class="editorHtml">fill=</span><span class="editorAttrVal">"green"</span><span class="editorHtml">/&gt;</span>
     			         </p>
     			         <p class="tab2 editorHtml">&lt;/svg&gt;</p>
     			         <p class="tab1 editorHtml">&lt;/body&gt;
     			    </div>
     			    </div>
                </section>
                
                <section>
        		     <h2>Step 14</h2>
        		     <br>
            	     <h3>Examine your test and reference files in the browser and confirm they are identical</h3>
            	     <img src="assets/images/passingRefTestExample.png">
                </section>
                <section>
        		     <h3>Failing test example</h3>
        		     <br>
            	     <img src="assets/images/failingRefTestExample.png">
                </section>
                <section>
                    <section>
    			        <h2>Step 15 - OSX</h2>
    			        <h5>Press the down-arrow for Windows instructions</h5>
    			        <br>
            		    <h3>Add your new test & reference file to the repository</h3>
          				<div class="console">
          				    <div class="consoleText">
              				    <br>
              				    <p>$ cd <span class="localPath">[path to repo]</span>/test/contributors/ttwf/<span class="localPath">yourName</span>/incoming</p>
              			        <p>$ hg add ttwf-reftest-tutorial-001.html</p>
              			        <p>$ hg add reftest</p>
              			        <p>$ hg commit -m "TTWF reftest tutorial"</p>
              			        <p>$ hg push</p>
              			    </div>
          			    </div>
                         <div class="bottomNote2">
          			            <p class="noteText">Note: If you receive the following error:</p> 
          			            <p class="errorText">abort: push creates new remote head ...</p>
          			            <p class="noteText">Your local repository is out of sync with the test repository</p>
          			            <p class="noteText">Do the following to resolve the issue:</p>
          			    </div>  
      			        <br><br><br><br><br><br><br>
                        <div class="console">
          				    <div class="consoleText">
              				    <br>
              			        <p>$ hg pull -u</p>
              			        <p>$ hg merge</p>
              			        <p>$ hg commit -m "Merge"</p>
              			        <p>$ hg push</p>
              			    </div>
          			    </div>          			    
      			    </section>
      			    <section>
    			        <h2>Step 15 - Windows</h2>
    			        <h5>Press the up-arrow for OSX instructions</h5>
    			        <br>
            		    <h3>Add your new test & reference file to the repository</h3>
          				<div class="console">
          				    <div class="dosText">
              				    <br>
              				    <p>> cd <span class="localPath">[path to repo]</span>\test\contributors\ttwf\<span class="localPath">yourName</span>\incoming</p>
              			        <p>$ hg add ttwf-reftest-tutorial-001.html</p>
              			        <p>$ hg add reftest</p>
              			        <p>> hg commit -m "TTWF reftest tutorial"</p>
              			        <p>> hg push</p>
              			    </div>
          			    </div>
                        <div class="bottomNote2">
      			            <p class="noteText">Note: If you receive the following error:</p> 
      			            <p class="errorText">abort: push creates new remote head ...</p>
      			            <p class="noteText">Your local repository is out of sync with the test repository</p>
      			            <p class="noteText">Do the following to resolve the issue:</p>
      			        </div>  
      			        <br><br><br><br><br><br><br>
                        <div class="console">
          				    <div class="dosText">
              				    <br>
              			        <p>> hg pull -u</p>
              			        <p>> hg merge</p>
              			        <p>> hg commit -m "Merge"</p>
              			        <p>> hg push</p>
              			    </div>
          			    </div>      			               			    
      			    </section>
                </section>
                <section>
    			     <h2>Step 16</h2>
    			     <br>
        		     <h3>Ask someone to review your test</h3>
    		    </section>
    		    <section>
    		        <section>
        			    <h2>Step 17 - OSX</h2>
        			    <h5>Press down-arrow for Windows instructions</h5>
        			    <br>
            		    <h3>Incorporate feedback & move test to the submitted folder</h3>
        			    <br>
          				<div class="consoleLarge">
          				    <div class="consoleText">
              				    <p>$ cd <span class="localPath">[path to repo]</span>/test/contributors/ttwf/<span class="localPath">yourName</span></p>
              				    <p>$ hg pull -u</p>
              				    <p>$ hg merge</p>
              				    <p>$ hg commit -m "merging"</p>
              			        <p>$ hg mv incoming/ttwf-reftest-tutorial-001.html submitted/ttwf-reftest-tutorial-001.html</p>
              			        <p>$ hg mv incoming/reftest/ttwf-reftest-tutorial-ref.html submitted/reftest/ttwf-reftest-tutorial-ref.html</p>
              			        <p>$ hg commit -m "moved the TTWF reftest tutorial to the submitted folder"</p>
              			        <p>$ hg push</p>
              			    </div>
          			    </div>
          		    </section>
                    <section>
    			        <h2>Step 17 - Windows</h2>
        			    <h5>Press down-arrow for OSX instructions</h5>
        			    <br>
            		    <h3>Incorporate feedback & move test to the submitted folder</h3>
        			    <br>
          				<div class="consoleLarge">
          				    <div class="dosText">
              				    <p>> cd <span class="localPath">[path to repo]</span>\test\contributors\ttwf\<span class="localPath">yourName</span></p>
              				    <p>> hg pull -u</p>
              				    <p>> hg merge</p>
              				    <p>> hg commit -m "merging"</p>
              			        <p>> hg mv incoming\ttwf-reftest-tutorial-001.html submitted\ttwf-reftest-tutorial-001.html</p>
              			        <p>> hg mv incoming\reftest/ttwf-reftest-tutorial-ref.html submitted\reftest\ttwf-reftest-tutorial-ref.html</p>
              			        <p>> hg commit -m "moved the TTWF reftest tutorial to the submitted folder"</p>
              			        <p>> hg push</p>
              			    </div>
          			    </div>
      		        </section>          		        
                </section>
                
                <section>
                    <h1>W3C JavaScript Test Tutorial</h1>
                </section>
                
				<section>
					<h2>JavaScript Test Overview</h2>
					<br>
					<ul>
					    <li>In most cases, JavaScript is the preferred method for writing tests</li>
					    <li>JavaScript tests have some advantages over Ref Tests
					        <ul>
					            <li>Robust</li>
					            <li>Flexible</li>
					            <li>Performant</li>
					            <li>Powerful</li>
					        </ul>
					    </li>
					    <li>W3C provides a framework (testharness) to simplify and standardize test creation
					        <ul>
					            <li>Includes an API for making common test assertions</li>
					            <li>Formats results and displays debug information for failures</li>
					        </ul>
					    </li>
					    <li>JavaScript tests are best when used in conjunction with some number of Ref Tests that verify final rendering</li>
					</ul>
				</section>
					
				<section>
				    <h2>JavaScript Test Tutorial</h2>
					<h3>Sample Test Case</h3>
					<br>
					<h4>Write a basic test for the SVG transform attribute described in the CSS3 Transforms spec using the translate() function</h4>
					<h4 style="color:red">Same as Ref Test example</h4>
					<br>
					<h3>Spec Links:</h3>
					<br>
				    <h4>
					<a href="http://dev.w3.org/csswg/css3-transforms/#svg-transform" target="top">http://dev.w3.org/csswg/css3-transforms/#svg-transform</a>
					<br>
					<br>
					<a href="http://dev.w3.org/csswg/css3-transforms/#two-d-transform-functions" target="top">http://dev.w3.org/csswg/css3-transforms/#two-d-transform-functions</a>
					</h4>
					
					<script>
						// Delicously hacky. Look away.
						if( navigator.userAgent.match( /(iPhone|iPad|iPod|Android)/i ) )
						document.write( '<p style="color: rgba(0,0,0,0.3); text-shadow: none;">('+'Tap to navigate'+')</p>' );
					</script>
				</section>
				                				
			    <section>
			        <h2>Step 1</h2>
			        <br>
    			    <h3>A JavaScript starter file was copied during Setup</h3>
    			    <h3>Open the <span style="color:red">ttwf-js-tutorial-001.html</span> file in an editor</h3>
    			    <img src="assets/images/jsTestStarter.png">
                </section>
                
                <section>
    			     <h2>Step 2</h2>
    			     <br>
        		     <h3>Review metadata</h3>
        		     <br>
        		     <h3>
    			        <ul>
    			            <li>Most items have been completed in the starter file</li>
    			            <li>Update the author metadata to include your name and email address</li>
    			            <li>Note the following differences from the Ref Test metadata:
    			                <ul>
    			                    <li>Additional flag was used: "dom"</li>
    			                    <li>No match metadata is required - only needed for Ref Tests</li>
    			                </ul>
    			            </li>
      				    </ul>
      				</h3>
                </section>
                
                <section>
        			<h2>Step 3</h2>
    			    <br>
        		    <h3>Import the W3C JavaScript Test Harness files</h3> 
        		    <br>
        		    <h4>Add the following lines to the starter file above the style block</h4>   
        		    <br>            
                    <div class="editorSmall">
                        <div class="editorText">
                            <br>
                            <p class="tab1 editorHtml">&lt;script src=<span class="editorAttrVal">"/resources/testharness.js"</span>&gt;&lt;/script&gt;</p>
                            <p class="tab1 editorHtml">&lt;script src=<span class="editorAttrVal">"/resources/testharnessreport.js"</span>&gt;&gt;&lt;/script&gt;</p>
                        </div>
                    </div>
                    <div class="bottomEditorNote">
                        <h5>Note: The path to the resources directory should match its location on your local machine</h5>
                        <br>
                        <h5>The path to these scripts must be updated to match the above before pushing to the test repo</h5>
                    </div>
                </section>

                <section>
        			<h2>Step 4</h2>
    			    <br>
        		    <h3>Add a div element with id="log" for test results</h3>
        		    <br>
                    <div class="editor">
       				    <div class="editorText">
           				    <br><br>
           				    <p class="tab1 editorHtml">&lt;body&gt;</p>
           				    <p class="tab2 editorHtml">...</p>
           			        <p class="tab2 editorHtml">&lt;div id=<span class="editorAttrVal">"log"</span>&gt;&lt;/div&gt;</span>
           			        <p class="tab1 editorHtml">&lt;/body&gt;
       			        </div>
       			     <br><br>
       			     <h4 style="line-height:30px">testharness will add an HTML table that contains a row for the test results from each test case in the file to this element</h4>
                </section>
                
                <section>
    			     <h2>Step 5</h2>
    			     <br>
        		     <h3>Set margins and padding to 0px</h3>
        		     <br>
        		     <div class="editor">
       			         <div class="editorText">
           			         <br><br>
           				     <p class="tab1 editorHtml">&lt;style&gt;</p>
           				     <p class="tab2 editorHtml">* {</p>
           			         <p class="tab3 editorHtml">margin: 0px;</p>
           			         <p class="tab3 editorHtml">padding: 0px;</p>
           			         <p class="tab2 editorHtml">}</p>
           			         <p class="tab1 editorHtml">&lt;/style&gt;
           			     </div>
       			     </div>
       			     <br><br><br><br><br><br>
       			     <h4>This is not required, but simplifies the code</h4>
    		    </section>    		    
    		    
                <section>
    			     <h2>Step 6</h2>
    			     <br>
        		     <h3>Create a script block that will contain the JavaScript</h3>
    			     <br>
         		     <div class="editor">
        			     <div class="editorText">
        			         <br><br>
                		     <p class="tab1 editorHtml">&lt;body&gt;</p>
           				     <p class="tab2 editorHtml">...</p>        			            
                		     <p class="tab2 editorHtml">&lt;script&gt;</p>
                		     <p class="tab2 editorHtml">&lt;/script&gt;</p>
                			 <p class="tab1 editorHtml">&lt;/body&gt;</p>
                		 </div>
        			 </div>
                </section>
                
                <section>
                    <h3>Test file should now appear as follows:</h3>
                    <img src="jsTest_step6.png">
                </section>
                
                <section>
    			     <h2>Step 7</h2>
    			     <br>
        		     <h3>Create a function inside the script block that will return the x,y position of an element</h3>
        		     <br>
        		     <div class="editor">
         			 <div class="editorText">
         			     <br>
         			     <p class="tab1 editorHtml">&lt;script&gt;
             		     <p class="tab2 editorHtml">function checkElementPosition(elemID) {</p>
             		     <p class="tab3 editorHtml"><span class="editorReserveWord">return</span> document.getElementById(elemID).getBoundingClientRect().top +</p>
             		     <p class="tab3 editorHtml">"," +  document.getElementById(elemID).getBoundingClientRect().left;</p>
             		     <p class="tab2 editorHtml">}</p>
             			 <p class="tab1 editorHtml">&lt;/script&gt;
             		 </div>
         			 </div>
    			</section>
    			
    			<section>
    			    <h2>The <span style="color:red">test</span> function</h2>
    			    <br>
    			    <ul>
    			        <li>Documented and implemented in testharness.js</li>
    			        <li>Function signature: <span style="color:red">test(test_function, name, properties)</span></li>
    			        <li>Used to define an individual test case in a file
    			            <ul>
    			                <li>Multiple calls to test() are common in a single file</li>
    			            </ul>
    			        </li>
    			        <li>The function argument must be a function object, not a function call</li>
    			        <li>The return value of the function object passed to test() (argument 1) determines Pass/Fail
    			            <ul>
    			                <li>The assert methods in testharness.js are typically used as the function supplied to test(), but can also be user defined</li>
    			            </ul>
    			        </li>
    			        <li>The name passed to test() (argument 2) should a be descriptive test name that conveys what is being tested
    			            <ul>
    			                <li>Name will be displayed in the test results table</li>
    			            </ul>
    			        </li>
    			        <li>Properties (argument 3) is an object that overrides default test properties
    			            <ul>
    			                <li>timeout is the only currently recognized property</li>
    			            </ul>
    			        </li>
    			    </ul>
    		    </section>
    		
                <section>
    			     <h2>Step 8</h2>
    			     <br>
        		     <h3>Add a call to the test function to your script block</h3>
        		     <h4 style="color:red">test(test_function, name, properties)</h4>
        		     <br>
        		     <div class="editor">
         			     <div class="editorText">
             			     <br><br>
             			     <p class="tab1 editorHtml">&lt;script&gt;
                 		     <p class="tab2 editorHtml">...</p>
                 		     <p class="tab2 editorHtml">test(function, "Test CSS transform translate on a div");</p>
                 			 <p class="tab1 editorHtml">&lt;/script&gt;
                 		 </div>
         			 </div>
         			 <br><br><br><br><br><br>
         			 <h4>No need to specify properties in this case</h4>
    			</section>    	
    			
    			<section>
    			    <h2>testharness.js assert methods</h2>
    			    <br>
    			    <ul>
    			        <li>testharness.js provides a number of assert methods to use for testing
    			            <ul>
    			                <li>Full list is provided in the comment section of the testharness.js file</li>
    			                <li>Whenever possible, these methods should be used</li>
    			            </ul>
    			        </li>
    			        <li>All assert methods contain an optional "description" argument that is only output if the assertion fails
    			            <ul>
    			                <li>The description will be added to the default output from testharness</li>
    			                <li>The description should be used to provide additional debug information in the case of a failure</li>
    			            </ul>
    			        </li>
    			        <li>All asserts must be located in a test()
    			            <ul>
    			                <li>asserts outside won't be detected correctly by the harness and may cause a file to stop testing</li>
    			            </ul>
    			    </ul>
    			</section>		
    			
                <section>
    			     <h2>Step 9</h2>
    			     <br>
        		     <h3>Add assert_equals as the test_function for test()</h3>
        		     <h4 style="color:red">assert_equals(actual, expected, description)</h4>
        		     <br>
        		     <div class="editorSmall">
           			     <div class="editorText">
                               <br>    			         
           			         <p class="tab2 editorHtml">test(function() {assert_equals(checkElementPosition(<span class="editorAttrVal">"myDiv"</span>), <span class="editorAttrVal">"50,50"</span>,</p>
           			         <p class="tab2 editorHtml"><span class="editorAttrVal">"myDiv should be at 50,50"</span>)}, <span class="editorAttrVal">"Test transform translate on SVG rect"</span>);</p>
               		     </div>
           			 </div>
         			 <br><br><br>
         			 <div style="text-align:left; line-height:20px; font-size:18px">
         			     <p>Arguments used for assert_equals:</p>
         			     <p>actual = assert_equals(checkElementPosition("myDiv")</p>
            		     <p>expected = "50,50"</p>
            		     <p>description = "myDiv should be at 50,50"</p>
            		</div>
    			</section>
                 
                <section>
                    <h2 style="line-height:10px">Final file</h2>
                    <img src="assets/images/jsTestFinal.png">
                </section>
                
                <section>
        		     <h2>Passing test example</h3>
        		     <br>
            	     <img src="assets/images/passingJSTestExample.png">
                </section>
                
                <section>
    			     <h2>Failing test example</h2>
    			     <br>
             	     <img src="assets/images/failingJSTestExample.png">
                </section>
                
                <section>
                    <section>
    			        <h2>Step 10 - OSX</h2>
    			        <h5>Press the down-arrow for Windows instructions</h5>
    			        <br>
            		    <h3>Add your new test file to the repository</h3>
          				<div class="console">
          				    <div class="consoleText">
              				    <br>
              				    <p>$ cd <span class="localPath">[path to repo]</span>/test/contributors/ttwf/<span class="localPath">yourName</span>/incoming</p>
              			        <p>$ hg add ttwf-js-tutorial-001.html</p>
              			        <p>$ hg commit -m "TTWF JavaScript tutorial"</p>
              			        <p>$ hg push</p>
              			    </div>
          			    </div>
                         <div class="bottomNote2">
          			            <p class="noteText">Note: If you receive the following error:</p> 
          			            <p class="errorText">abort: push creates new remote head ...</p>
          			            <p class="noteText">Your local repository is out of sync with the test repository</p>
          			            <p class="noteText">Do the following to resolve the issue:</p>
          			    </div>  
      			        <br><br><br><br><br><br><br>
                        <div class="console">
          				    <div class="consoleText">
              				    <br>
              			        <p>$ hg pull -u</p>
              			        <p>$ hg merge</p>
              			        <p>$ hg commit -m "Merge"</p>
              			        <p>$ hg push</p>
              			    </div>
          			    </div>          			       
      			    </section>
      			    <section>
    			        <h2>Step 10 - Windows</h2>
    			        <h5>Press the up-arrow for OSX instructions</h5>
    			        <br>
            		    <h3>Add your new test file to the repository</h3>
          				<div class="console">
          				    <div class="dosText">
              				    <br>
              				    <p>> cd <span class="localPath">[path to repo]</span>\test\contributors\ttwf\<span class="localPath">yourName</span>\incoming</p>
              			        <p>> hg add ttwf-js-tutorial-001.html</p>
              			        <p>> hg commit -m "TTWF JavaScript tutorial"</p>
              			        <p>> hg push</p>
              			    </div>
          			    </div>
                        <div class="bottomNote2">
      			            <p class="noteText">Note: If you receive the following error:</p> 
      			            <p class="errorText">abort: push creates new remote head ...</p>
      			            <p class="noteText">Your local repository is out of sync with the test repository</p>
      			            <p class="noteText">Do the following to resolve the issue:</p>
      			        </div>  
      			        <br><br><br><br><br><br><br>
                        <div class="console">
          				    <div class="dosText">
              				    <br>
              			        <p>> hg pull -u</p>
              			        <p>> hg merge</p>
              			        <p>> hg commit -m "Merge"</p>
              			        <p>> hg push</p>
              			    </div>
          			    </div>      			               			    
      			    </section>
                </section>
                
                <section>
    			     <h2>Step 11</h2>
    			     <br>
        		     <h3>Ask someone to review your test</h3>
    		    </section>
    		    
    		    <section>
    		        <section>
        			    <h2>Step 12 - OSX</h2>
        			    <h5>Press down-arrow for Windows instructions</h5>
        			    <br>
            		    <h3>Incorporate feedback & move test to the submitted folder</h3>
        			    <br>
          				<div class="consoleLarge">
          				    <div class="consoleText">
          				        <br><br>
              				    <p>$ cd <span class="localPath">[path to repo]</span>/test/contributors/ttwf/<span class="localPath">yourName</span></p>
              			        <p>$ hg mv incoming/ttwf-js-tutorial-001.html submitted/ttwf-js-tutorial-001.html</p>
              			        <p>$ hg commit -m "moved the TTWF JavaScript tutorial to the submitted folder"</p>
              			        <p>$ hg push</p>
              			    </div>
          			    </div>
          		    </section>
                    <section>
    			        <h2>Step 12 - Windows</h2>
        			    <h5>Press down-arrow for OSX instructions</h5>
        			    <br>
            		    <h3>Incorporate feedback & move test to the submitted folder</h3>
        			    <br>
          				<div class="consoleLarge">
          				    <div class="dosText">
          				        <br><br>
              				    <p>> cd <span class="localPath">[path to repo]</span>\test\contributors\ttwf\<span class="localPath">yourName</span></p>
              			        <p>> hg mv incoming\ttwf-js-tutorial-001.html submitted\ttwf-js-tutorial-001.html</p>
              			        <p>> hg commit -m "moved the TTWF JavaScript tutorial to the submitted folder"</p>
              			        <p>> hg push</p>
              			    </div>
          			    </div>
      		        </section>          		        
                </section>  
                
                <section>
                    <h2>Congratulations!</h2>
                    <br><br>
                    <h3>You have completed your first W3C tests!</h3>
                    <br><br>
                    <h3>TODO: Add funny image here?</h3>
                </section>                 		    		     
        </div>

			<!-- The navigational controls UI -->
			<aside class="controls">
				<a class="left" href="#">&#x25C4;</a>
				<a class="right" href="#">&#x25BA;</a>
				<a class="up" href="#">&#x25B2;</a>
				<a class="down" href="#">&#x25BC;</a>
			</aside>

			<!-- Displays presentation progress, max value changes via JS to reflect # of slides -->
			<div class="progress"><span></span></div>
			
		</div>
		
		<script src="js/reveal.js"></script>
		<script src="js/prefixfree.min.js"></script>
		<script src="lib/highlight.js"></script>
		<script>
			// Parse the query string into a key/value object
			var query = {};
			location.search.replace( /[A-Z0-9]+?=(\w*)/gi, function(a) {
				query[ a.split( '=' ).shift() ] = a.split( '=' ).pop();
			} );

			Reveal.initialize({
				// Display controls in the bottom right corner
				controls:true,

				// Display a presentation progress bar
				progress: true,

				// If true; each slide will be pushed to the browser history
				history: true,

				// Flags if mouse wheel navigation should be enabled
				mouseWheel: true,

				// Apply a 3D roll to links on hover
				rollingLinks: true,

				// UI style
				theme: query.theme || 'default', // default/neon

				// Transition style
				transition: query.transition || 'concave' // default/cube/page/concave/linear(2d)
			});

			hljs.initHighlightingOnLoad();
		</script>

	</body>
</html>
